+++ modules/stempel/src/lmgrep/stempel.clj
@@ -1,5 +1,0 @@
-(ns lmgrep.stempel
-  (:import (org.apache.lucene.analysis.pl PolishAnalyzer)))
-
-(def analyzers
-  {"polish" (PolishAnalyzer.)})